Remove token-based Input in favor of character-based Text as a step towards collapsing the suspiciously similar lexer/parser models.
As of this commit, the solution compiles but does not pass all tests.
The changes in this commit are the result of deleting Input, replacing usages of Input with Text, and performing only trivial superficial fixups from there.
Note that OperatorPrecedenceParser itself poses an extra challenge, so rather than performing its fixes yet it is left in a not-implemented state to be resolved in subsequent commits.